Removed references to backend-get-routes.js. Added it's contents to api-routes.js.
