+++ server.js
@@ -45,7 +45,6 @@ require("./routes/business_api_routes.js")(app);
 require("./routes/comment_api_routes.js")(app);
 require("./routes/customer_api_route.js")(app);
 require("./routes/review_api_route.js")(app);
-require("./routes/backend-get-routes.js")(app);
 
 // Syncing our sequelize models and then starting our Express app
 // =============================================================
